+++ packages/react-tinacms-inline/src/index.tsx
@@ -18,9 +18,20 @@ limitations under the License.
 
 export * from './inline-form'
 export * from './inline-field'
-export * from './inline-field-text'
-export * from './inline-field-textarea'
-export * from './inline-field-image'
-export * from './inline-wysiwyg'
+export {
+  InlineText,
+  InlineTextField,
+  InlineTextProps,
+} from './fields/inline-text-field'
+export {
+  InlineTextarea,
+  InlineTextareaField,
+} from './fields/inline-textarea-field'
+export {
+  InlineImage,
+  InlineImageField,
+  InlineImageProps,
+} from './fields/inline-image-field'
+export * from './inline-group'
 export * from './blocks'
 export * from './styles'